wip: layernorm refactoring in progress
